simple server

+++ storage_service/database.py
@@ -2,8 +2,8 @@
 from typing import Optional
 from abc import ABC, abstractmethod
 
-import dataclasses
-from pathlib import Path
+import dataclasses, hashlib
+from enum import IntEnum
 
 import asyncio
 import aiosqlite
@@ -11,13 +11,16 @@ import aiofiles
 import aiofiles.os
 import aiofiles.ospath
 
-from .config import DATA_HOME
+from .config import DATA_HOME, FILE_ROOT
 from .log import get_logger
 
 _g_conn: Optional[aiosqlite.Connection] = None
 
+def hash_credential(username, password):
+    return hashlib.sha256((username + password).encode()).hexdigest()
+
 class DBConnBase(ABC):
-    logger = get_logger('database', buffer_size=128, global_instance=True)
+    logger = get_logger('database', global_instance=True)
 
     @property
     def conn(self)->aiosqlite.Connection:
@@ -40,7 +43,7 @@ class DBConnBase(ABC):
 class DBUserRecord:
     id: int
     username: str
-    password: str
+    credential: str
     is_admin: bool
     create_time: str
     last_active: str
@@ -52,7 +55,7 @@ DECOY_USER = DBUserRecord(0, 'decoy', 'decoy', False, '2021-01-01 00:00:00', '20
 class UserConn(DBConnBase):
 
     @staticmethod
-    def parse_record(record: list) -> DBUserRecord:
+    def parse_record(record) -> DBUserRecord:
         return DBUserRecord(*record)
 
     async def init(self):
@@ -61,7 +64,7 @@ class UserConn(DBConnBase):
         CREATE TABLE IF NOT EXISTS user (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             username VARCHAR(255) UNIQUE NOT NULL,
-            password VARCHAR(255) NOT NULL,
+            credential VARCHAR(255) NOT NULL,
             is_admin BOOLEAN DEFAULT FALSE,
             create_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP, 
             last_active TIMESTAMP DEFAULT CURRENT_TIMESTAMP
@@ -83,15 +86,38 @@ class UserConn(DBConnBase):
         if res is None: return None
         return self.parse_record(res)
     
+    async def get_user_by_credential(self, credential: str) -> Optional[DBUserRecord]:
+        async with self.conn.execute("SELECT * FROM user WHERE credential = ?", (credential, )) as cursor:
+            res = await cursor.fetchone()
+        
+        if res is None: return None
+        return self.parse_record(res)
+    
     async def create_user(self, username: str, password: str, is_admin: bool = False) -> int:
         self.logger.debug(f"Creating user {username}")
+        credential = hash_credential(username, password)
         assert await self.get_user(username) is None
-        async with self.conn.execute("INSERT INTO user (username, password, is_admin) VALUES (?, ?, ?)", (username, password, is_admin)) as cursor:
+        async with self.conn.execute("INSERT INTO user (username, credential, is_admin) VALUES (?, ?, ?)", (username, credential, is_admin)) as cursor:
             self.logger.info(f"User {username} created")
+            assert cursor.lastrowid is not None
             return cursor.lastrowid
     
-    async def set_user(self, username: str, password: str, is_admin: bool = False):
-        await self.conn.execute("UPDATE user SET password = ?, is_admin = ? WHERE username = ?", (password, is_admin, username))
+    async def set_user(self, username: str, password: Optional[str] = None, is_admin: Optional[bool] = None):
+        if password is not None:
+            credential = hash_credential(username, password)
+        else:
+            async with self.conn.execute("SELECT credential FROM user WHERE username = ?", (username, )) as cursor:
+                res = await cursor.fetchone()
+                assert res is not None, f"User {username} not found"
+                credential = res[0]
+        
+        if is_admin is None:
+            async with self.conn.execute("SELECT is_admin FROM user WHERE username = ?", (username, )) as cursor:
+                res = await cursor.fetchone()
+                assert res is not None, f"User {username} not found"
+                is_admin = res[0]
+
+        await self.conn.execute("UPDATE user SET credential = ?, is_admin = ? WHERE username = ?", (credential, is_admin, username))
         self.logger.info(f"User {username} updated")
     
     async def all(self):
@@ -106,21 +132,27 @@ class UserConn(DBConnBase):
         await self.conn.execute("DELETE FROM user WHERE username = ?", (username, ))
         self.logger.info(f"Delete user {username}")
 
+class FileReadPermission(IntEnum):
+    PUBLIC = 0          # accessible by anyone
+    PROTECTED = 1       # accessible by any user
+    PRIVATE = 2         # accessible by owner only (including admin)
 
 @dataclasses.dataclass
 class FileDBRecord:
     url: str
     user_id: int
-    file_path: str
+    file_path: str      # relative to FILE_ROOT
+    file_size: int
     create_time: str
+    permission: FileReadPermission
 
     def __str__(self):
-        return f"File {self.url} (user={self.user_id}, created at {self.create_time}, path={self.file_path})"
+        return f"File {self.url} (user={self.user_id}, created at {self.create_time}, path={self.file_path}, permission={self.permission}, size={self.file_size})"
     
 class FileConn(DBConnBase):
 
     @staticmethod
-    def parse_record(record: list) -> FileDBRecord:
+    def parse_record(record) -> FileDBRecord:
         return FileDBRecord(*record)
     
     async def init(self):
@@ -131,7 +163,8 @@ class FileConn(DBConnBase):
             user_id INTEGER NOT NULL,
             file_path VARCHAR(255) NOT NULL,
             file_size INTEGER,
-            create_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+            create_time TIMESTAMP DEFAULT CURRENT_TIMESTAMP, 
+            permission INTEGER DEFAULT 0
         )
         ''')
         return self
@@ -153,18 +186,25 @@ class FileConn(DBConnBase):
             res = await cursor.fetchall()
         return [self.parse_record(r) for r in res]
     
-    async def set_file_record(self, url: str, user_id: int, file_path: str):
+    async def set_file_record(self, url: str, user_id: int, file_path: str, permission: Optional[ FileReadPermission ] = None):
+        """file_path is relative to FILE_ROOT"""
         self.logger.debug(f"Updating file {url}: user_id={user_id}, file_path={file_path}")
 
-        assert await aiofiles.ospath.exists(file_path), f"File {file_path} not found"
-        file_size = (await aiofiles.os.stat(file_path)).st_size
+        abs_file_path = FILE_ROOT / file_path
+        assert await aiofiles.ospath.exists(abs_file_path), f"File {abs_file_path} not found"
+        file_size = (await aiofiles.os.stat(abs_file_path)).st_size
         
         old = await self.get_file_record(url)
         if old is not None:
-            await self.conn.execute("UPDATE file SET user_id = ?, file_path = ?, file_size = ? WHERE url = ?", (user_id, file_path, file_size, url))
+            assert old.user_id == user_id, f"User mismatch: {old.user_id} != {user_id}"
+            if permission is None:
+                permission = old.permission
+            await self.conn.execute("UPDATE file SET file_path = ?, file_size = ?, permission = ? WHERE url = ?", (file_path, file_size, permission, url))
             self.logger.info(f"File {url} updated")
         else:
-            await self.conn.execute("INSERT INTO file (url, user_id, file_path, file_size) VALUES (?, ?, ?, ?)", (url, user_id, file_path, file_size))
+            if permission is None:
+                permission = FileReadPermission.PUBLIC
+            await self.conn.execute("INSERT INTO file (url, user_id, file_path, file_size, permission) VALUES (?, ?, ?, ?, ?)", (url, user_id, file_path, file_size, permission))
             self.logger.info(f"File {url} created")
     
     async def delete_file_record(self, url: str):
@@ -177,13 +217,13 @@ class FileConn(DBConnBase):
         async with self.conn.execute("SELECT * FROM file WHERE user_id = ?", (user_id, )) as cursor:
             res = await cursor.fetchall()
         await self.conn.execute("DELETE FROM file WHERE user_id = ?", (user_id, ))
-        self.logger.info(f"Deleted {len(res)} files for user {user_id}")
+        self.logger.info(f"Deleted {len(res)} files for user {user_id}") # type: ignore
     
     async def delete_path_records(self, path: str):
         async with self.conn.execute("SELECT * FROM file WHERE file_path LIKE ?", (path + '%', )) as cursor:
             res = await cursor.fetchall()
         await self.conn.execute("DELETE FROM file WHERE file_path LIKE ?", (path + '%', ))
-        self.logger.info(f"Deleted {len(res)} files for path {path}")
+        self.logger.info(f"Deleted {len(res)} files for path {path}") # type: ignore
 
 async def _remove_files_if_exist(files: list):
     async def remove_file(file_path):
@@ -192,7 +232,7 @@ async def _remove_files_if_exist(files: list):
     await asyncio.gather(*[remove_file(f) for f in files])
 
 def _validate_url(url: str) -> bool:
-    return url.startswith('/') and not ('..' in url)
+    return not url.startswith('/') and not ('..' in url) and ('/' in url)
 
 async def get_user(db: "Database", user: int | str) -> Optional[DBUserRecord]:
     if isinstance(user, str):
@@ -202,7 +242,6 @@ async def get_user(db: "Database", user: int | str) -> Optional[DBUserRecord]:
     else:
         return None
 
-FILE_ROOT = DATA_HOME / 'files'
 @dataclasses.dataclass(frozen=True)
 class Database:
     user: UserConn = UserConn()
@@ -223,7 +262,7 @@ class Database:
         if _g_conn is not None:
             await _g_conn.close()
 
-    async def save_file(self, u: int | str, url: str, blob: bytes):
+    async def save_file(self, u: int | str, url: str, blob: bytes | str):
         if not _validate_url(url):
             raise ValueError(f"Invalid URL: {url}")
 
@@ -232,24 +271,26 @@ class Database:
             return
 
         username = user.username
-        assert url.startswith("/" + username), f"URL must start with /{username}, get: {url}"
+        assert url.startswith(f"{username}/"), f"URL must start with {username}/, get: {url}"
 
         file_path = FILE_ROOT / url
         await aiofiles.os.makedirs(file_path.parent, exist_ok=True)
+
+        if isinstance(blob, str):
+            blob = blob.encode("utf-8")
         async with aiofiles.open(file_path, 'wb') as f:
             await f.write(blob)
 
         rel_path = str(file_path.relative_to(FILE_ROOT))
-        if not rel_path.startswith('/'):
-            rel_path = '/' + rel_path
+        assert _validate_url(rel_path)
 
-        await self.file.set_file_record(url, user, str(rel_path))
+        await self.file.set_file_record(url, user.id, str(rel_path))
         await self.user.set_active(username)
         await self.commit()
 
         return rel_path
 
-    async def get_file(self, url: str) -> bytes:
+    async def read_file(self, url: str) -> bytes:
         if not _validate_url(url): raise ValueError(f"Invalid URL: {url}")
 
         r = await self.file.get_file_record(url)
@@ -257,17 +298,28 @@ class Database:
             raise FileNotFoundError(f"File {url} not found")
         async with aiofiles.open(FILE_ROOT / r.file_path, 'rb') as f:
             return await f.read()
-
-    async def delete_file(self, url: str):
+    
+    async def read_file_stream(self, url: str):
         if not _validate_url(url): raise ValueError(f"Invalid URL: {url}")
 
         r = await self.file.get_file_record(url)
         if r is None:
-            return
-        await aiofiles.os.remove(FILE_ROOT / r.file_path)
+            raise FileNotFoundError(f"File {url} not found")
+        fpath = FILE_ROOT / r.file_path
+        fsize = (await aiofiles.os.stat(fpath)).st_size
+        return (fsize, await aiofiles.open(fpath, 'rb'))
+
+    async def delete_file(self, url: str) -> Optional[FileDBRecord]:
+        if not _validate_url(url): raise ValueError(f"Invalid URL: {url}")
+
+        r = await self.file.get_file_record(url)
+        if r is None:
+            return None
+        if await aiofiles.ospath.exists(FILE_ROOT / r.file_path):
+            await aiofiles.os.remove(FILE_ROOT / r.file_path)
         await self.file.delete_file_record(url)
         await self.commit()
-        return
+        return r
 
     async def delete_files(self, url: str):
         if not _validate_url(url): raise ValueError(f"Invalid URL: {url}")
